Added basic layout for index.html and added files for formatting and ajax
I added a number of <div> tags for formatting the page.  The layout of the page is based on a grid; with Bootstraps rows (.row) acting as vertical separators, and columns (.col) acting as horizontal separators. I also added a js file for performing the AJAX operations in the program. More work will be done as components come in.